Fixed handling of values of QKeyEvent/key() larger than supported by uint16.

+++ src/vsgQt/KeyboardMap.h
@@ -15,7 +15,7 @@ namespace vsgQt
         bool getKeySymbol(const QKeyEvent* e, vsg::KeySymbol& keySymbol, vsg::KeySymbol& modifiedKeySymbol, vsg::KeyModifier& keyModifier);
 
     protected:
-        using VirtualKeyToKeySymbolMap = std::map<uint16_t, vsg::KeySymbol>;
+        using VirtualKeyToKeySymbolMap = std::map<uint32_t, vsg::KeySymbol>;
         VirtualKeyToKeySymbolMap _keycodeMap;
     };
